Updated Launch Page
Simplfied the loading screen
