+++ scouting/assets/images/desktop.ini
@@ -1,0 +1,2 @@
+[LocalizedFileNames]
+cute-cyberlion.png=@cute-cyberlion,0
