+++ scouting/package.json
@@ -13,7 +13,7 @@
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
     "react": "18.2.0",
-    "react-native": "0.72.3",
+    "react-native": "^0.72.3",
     "react-native-safe-area-context": "^4.7.1",
     "react-native-screens": "^3.24.0"
   },
